# Conflicts:
#	Front/systock/package-lock.json
#	Front/systock/package.json
#	Server/prisma/ERD.svg
#	Server/prisma/seed.ts
#	Server/src/controllers/SupplierController.ts
#	Server/src/middleware/auth_middleware.ts
#	Server/src/models/Supplier.ts
#	Server/src/routes.ts

+++ Front/systock/src/hooks/useCategory.js
@@ -1,7 +1,10 @@
 /* eslint-disable no-debugger */
-import { useEffect, useRef, useState } from "react";
+import { useContext, useEffect, useRef, useState } from "react";
 import { deepCopy } from "../utils/utils";
 import { performFetch, performFetchNoResult } from "../apiBase";
+import { DEBUG_LOCAL, MainContext } from "../App";
+import { products } from "../utils/data";
+import { ENTITIES } from "../utils/debug-local-helper";
 
 export default function useCategory() {
 
@@ -19,6 +22,8 @@ export default function useCategory() {
   const [autoHideSnackBar, setAutoHideSnackBar] = useState(3000);
   const [severitySnackBar, setSeveritySnackBar] = useState("info");
   const [snackMessageSnackBar, setSnackMessageSnackBar] = useState("");
+
+  const { updateData, getData } = useContext(MainContext);
 
   function handleOpenSnackBar(severity, message="Unexpected Error Occurred", autoHide=3000) {
     setSnackMessageSnackBar(message);
@@ -46,8 +51,13 @@ export default function useCategory() {
   }
 
   async function getCategories() {
+    if(DEBUG_LOCAL) {
+      const categArr = getData(ENTITIES.CATEGORIES);
+
+      return setTimeout(() => setCategories(categArr), 350);
+    }
     try {
-      const categories = await performFetch("/categories", {method: 'GET'});
+      const { categories } = await performFetch("/categories", {method: 'GET'});
       setCategories(categories);
     } catch (error) {
       console.log(error.message);
@@ -71,13 +81,18 @@ export default function useCategory() {
    * @param {*} category 
    */
   const insertCategory = (newCategory) => {
+
     const newItem = deepCopy(newCategory);
 
     setCategories(prevState => {
       if(prevState.some(cat => cat.id === newItem.id)) {
-        return prevState.map(cat => (cat.id === newItem.id ? {...newItem} : {...cat}));
+        const nextArr = prevState.map(cat => (cat.id === newItem.id ? {...newItem} : {...cat}));
+        if(DEBUG_LOCAL) updateData(ENTITIES.CATEGORIES, nextArr);
+        return nextArr;
       } else {
-        return [...prevState, newItem];
+        const nextArr = [...prevState, newItem];
+        if(DEBUG_LOCAL) updateData(ENTITIES.CATEGORIES, nextArr);
+        return nextArr;
       }
     });
     
@@ -89,6 +104,8 @@ export default function useCategory() {
    * @param {*} 
    */
   async function handleCreateCategory(obj) {
+    if(DEBUG_LOCAL) return insertCategory(obj);     
+
     try {
       const newItem = await performFetch("/category", {method: 'POST', body: JSON.stringify(obj)});
       if(typeof newItem === 'string') {
@@ -108,6 +125,7 @@ export default function useCategory() {
    * @param {*} 
    */
   async function handleUpdateCategory(category) {
+    if(DEBUG_LOCAL) return insertCategory(category);     
     try {
       const newItem = await performFetch("/category", {method: 'PUT', body: JSON.stringify(category)});
       
@@ -129,6 +147,11 @@ export default function useCategory() {
    * @param {*} id 
    */
   async function handleDeleteCategory(id) {
+    if(DEBUG_LOCAL) {
+      const updatedCategories = categories.filter(cat => cat.id !== id.id);
+      setCategories(updatedCategories);
+      return updateData(ENTITIES.CATEGORIES, updatedCategories);
+    }
     const url = "/category";
 
     performFetchNoResult(url, {method: 'DELETE', body: JSON.stringify(id)})
